[LEMS-2737/default-grapher] untangle grapher
